@@ -4,7 +4,7 @@ version = "0.8.0"
 description = "Full-featured async (tokio-based) connection pool (like r2d2)"
 license = "MIT"
 repository = "https://github.com/djc/bb8"
-edition = "2018"
+edition = "2021"
 workspace = ".."
 readme = "../README.md"
 
