@@ -4,7 +4,7 @@ version = "0.8.1"
 description = "Full-featured async (tokio-based) postgres connection pool (like r2d2)"
 license = "MIT"
 repository = "https://github.com/djc/bb8"
-edition = "2018"
+edition = "2021"
 
 [features]
 "with-bit-vec-0_6" = ["tokio-postgres/with-bit-vec-0_6"]
