@@ -4,7 +4,7 @@ version = "0.11.0"
 description = "Full-featured async (tokio-based) redis connection pool (like r2d2)"
 license = "MIT"
 repository = "https://github.com/djc/bb8"
-edition = "2018"
+edition = "2021"
 
 [dependencies]
 async-trait = "0.1"
